Changed authentication plugin to use Authlogic instead of restful_authentication.
This makes it possible for people to integrate Refinery into apps that
are already using Authlogic for authentication.

For those who were using Refinery's default (restful_authentication),
this change shouldn't affect you either. Authlogic's
act_like_restful_authentication = true option means that users should
still be able to log in even though their passwords were encrypted with
restful_authentication. The transition_from_restful_authentication
option means that "every time a user successfully logs in and their
password is encrypted with the restful_authentication algorithm it will
update their password with the Authlogic algorithm. When a new account
is created it will use the Authlogic algorithm. This allows your user
base to slowly transition and allowing them to still be able to log in."
(http://www.binarylogic.com/2008/11/23/tutorial-easily-migrate-from-restful_authentication-to-authlogic/)

The User model is also a lot simpler now, since most of the authentication
code is handled in the Authlogic gem itself. This leaves the model so
that it only needs to worry about application-specific stuff.