Fixed broken help links on fields with errors.
Rails wraps a invalid field (and its label) with a div of class
fieldWithErrors.  Since div is a block element, it causes the help link
to move to the next line.  The fix is to tell Rails to use a span
instead of a div for wrapping an invalid field.

For example:
old:
 <span class="label_with_help">
   <div class="fieldWithErrors">
     <label for="event_pricing_options_attributes_1_cost">Cost</label>
   </div>
   // Flow is broken.  div above and span below are on different lines.
   <span class="help" tooltip="Number only, no $."> (help) </span>
 </span>

new:
<span class="label_with_help">
  <span class="fieldWithErrors">
    <label for="event_pricing_options_attributes_1_cost">Cost</label>
  </span>
   // Flow is ok.  span above and span below are on same line.
  <span class="help" tooltip="Number only, no $."> (help) </span>
</span>