Migrations combined into one consistent migration which uses Model.table_name instead of hardcoded table names so that you could include RefineryCMS into an existing application that uses the same table names that we do (e.g. pages is a likely candidate).
Extracted Inquiries migration logic into the Inquiries engine which is what we should have done at 0.9.8 but didn't.

Moved Admin::BaseController into the core so that it always exists so that we could copy it into existing applications hooking into Refinery using an override rather than it having a persistent presence.

All of the migration logic assumes that you were up to date with 0.9.8.5