revised much of Heroku hosting guide
based on my personal experience, uploading an existing app to Heroku while attempting to follow the guide

I deleted the part about adding the Heroku gem because you should have already installed the tool using the quickstart, and the app itself doesn’t rely on the gem. I know that because my app works on Heroku without 'heroku' in the Gemfile.