Fix preview of image on insert
Problem: when inserting an image with the image picker in an image field, the preview is broken (broken image tag). (otherwise the preview works)

Solution: the image path is: `<path>?sha=some_sha?23432`, notice the random integer at the end, which isn't needed anymore (I assume). So I just deleted that part in the js file.

Background:
* the problem was mentioned by @eshaiju in https://github.com/refinery/refinerycms/issues/2681#issuecomment-56522971
* the random int was added here: https://github.com/refinery/refinerycms/commit/ec3f2a907116a8c1c9afc59bb082cdb8c3b9c2b5